lifecycle: add error handlers
Following on the interceptor inpiration, add support for an
`:error` key in steps. Error handlers which will receive the
step which threw, the current context and the exception.

If the handler yields a non-throwable value, let processing
continue, otherwise, walk up the chain of handler to try
finding other error handlers